fix Login styling and rename
