+++ client/src/pages/Login/RegisterForm.tsx
@@ -8,7 +8,7 @@ const RegisterForm = ({ user, errors, handleChange, handleSubmit }) => {
   const labelStyle = 'block text-sm mb-2';
 
   return (
-    <section className="bg-gray-50 dark:bg-gray-900 rounded-lg py-2">
+    <section className="bg-background dark:bg-gray-900 rounded-lg py-2">
       <div className="flex flex-col items-center px-6 py-8 mx-auto">
         <div className="w-full bg-background rounded-lg shadow-lg dark:border md:mt-0 sm:max-w-md xl:p-0 dark:bg-gray-800 dark:border-gray-700 bg-gradient-to-t from-primary-50 to-primary-10">
           <div className="p-6 space-y-4 md:space-y-6 sm:p-8 rounded-lg">
