+++ client/src/routes/Main.tsx
@@ -4,7 +4,7 @@ import AnimatedPage from '../components/AnimatedPage';
 import Home from '../pages/Home';
 import About from '@/pages/About';
 import NotFound from '../pages/NotFound';
-import Login from '../pages/Login/Login';
+import Login from '../pages/Login';
 import Register from '../pages/Login/Register';
 
 import ProtectedRoute from './ProtectedRoute';
